+++ debugger/src/dap_types.rs
@@ -2,6 +2,7 @@
 use crate::DebuggerError;
 use num_traits::Num;
 use parse_int::parse;
+use probe_rs_cli_util::rtt;
 use schemafy::schemafy;
 use serde::{Deserialize, Serialize};
 use std::convert::TryFrom;
@@ -29,7 +30,7 @@ pub struct QuitRequest {
 pub struct RttChannelEventBody {
     pub channel_number: usize,
     pub channel_name: String,
-    pub data_format: crate::rtt::DataFormat,
+    pub data_format: rtt::DataFormat,
 }
 
 #[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
