+++ probe-rs-cli-util/Cargo.toml
@@ -1,7 +1,10 @@
 [package]
 name = "probe-rs-cli-util"
 version = "0.12.0"
-authors = ["Noah Hüsser <yatekii@yatekii.ch", "Dominik Boehi <dominik.boehi@gmail.com>"]
+authors = [
+    "Noah Hüsser <yatekii@yatekii.ch",
+    "Dominik Boehi <dominik.boehi@gmail.com>",
+]
 edition = "2021"
 description = "Helper library for CLI applications based on probe-rs."
 documentation = "https://docs.rs/probe-rs-cli-util/"
@@ -16,6 +19,9 @@ license = "MIT OR Apache-2.0"
 default = ["anyhow"]
 
 [dependencies]
+probe-rs-rtt = { path = "../rtt" }
+probe-rs = { version = "0.12.0", path = "../probe-rs" }
+
 thiserror = "1.0"
 anyhow = { version = "1.0", optional = true }
 indicatif = "0.16.0"
@@ -23,9 +29,8 @@ env_logger = "0.9.0"
 log = "0.4.0"
 once_cell = "1.7.2"
 colored = "2.0.0"
-probe-rs = { version = "0.12.0", path = "../probe-rs" }
 cargo_toml = "0.11.1"
-serde = { version = "1.0.115", features = [ "derive" ] }
+serde = { version = "1.0.115", features = ["derive"] }
 cargo_metadata = "0.14.0"
 dunce = "1.0.1"
 sentry = { version = "0.23.0", features = ["anyhow"], optional = true }
@@ -33,3 +38,7 @@ simplelog = "0.11.0"
 terminal_size = "0.1.13"
 clap = { version = "3.0", features = ["derive"] }
 byte-unit = "4.0.13"
+chrono = { version = "0.4", features = ["serde"] }
+goblin = "0.4.1"
+num-traits = "0.2.14"
+defmt-decoder = { version = "0.2.1", features = ["unstable"] }
